+++ test/unit/ExceptionHandling/RequestExceptionDispatcherTest.php
@@ -10,12 +10,21 @@ namespace test\unit\Ingenerator\KohanaExtras\ExceptionHandling;
 use Ingenerator\KohanaExtras\DependencyContainer\DependencyContainer;
 use Ingenerator\KohanaExtras\ExceptionHandling\ExceptionHandler;
 use Ingenerator\KohanaExtras\ExceptionHandling\RequestExceptionDispatcher;
+use Ingenerator\PHPUtils\Logging\StackdriverApplicationLogger;
+use Ingenerator\PHPUtils\StringEncoding\JSON;
+use Psr\Log\LogLevel;
+use Psr\Log\NullLogger;
+use Psr\Log\Test\TestLogger;
 use Symfony\Component\Process\Process;
 
 class RequestExceptionDispatcherTest extends \PHPUnit\Framework\TestCase
 {
     protected $default_handler;
+
     protected $dependencies;
+
+    protected $logger;
+
     protected $handler_map = [];
 
     public function test_it_is_initialisable_when_kohana_is_all_initialised()
@@ -64,46 +73,88 @@ class RequestExceptionDispatcherTest extends \PHPUnit\Framework\TestCase
     public function test_its_static_constructor_gets_basic_working_instance_when_container_not_initialised_or_invalid(
         $args
     ) {
-        $dispatcher = RequestExceptionDispatcher::factory($args['class'], $args['method'], $args['service']);
-        $this->assertInstanceOf(RequestExceptionDispatcher::class, $dispatcher);
+        $this->withFakeFallbackLogger(
+            function () use ($args) {
+                $dispatcher = RequestExceptionDispatcher::factory($args['class'], $args['method'], $args['service']);
+                $this->assertInstanceOf(RequestExceptionDispatcher::class, $dispatcher);
+            }
+        );
     }
 
     public function test_its_static_constructor_creates_instance_that_basically_works_even_without_kohana()
     {
-        $file   = \realpath(__DIR__.'/../../../src/ExceptionHandling/RequestExceptionDispatcher.php');
-        $class  = RequestExceptionDispatcher::class;
-        $script = \implode(
-            "\n",
+        $basedir = \realpath(__DIR__.'/../../../');
+        $proc    = $this->runTemporaryPhpScript(
+            <<<'PHP'
+                use Ingenerator\KohanaExtras\ExceptionHandling\RequestExceptionDispatcher;
+                require REQUEST_EXCEPTION_DISPATCHER_SOURCE;
+                try {
+                  ob_start();
+                  echo "This should not make it to output";
+                  throw new \InvalidArgumentException("Testing handling exceptions");
+                } catch (\Exception $e) {
+                  RequestExceptionDispatcher::dispatch($e);
+                }
+                fwrite(STDERR, "Code: ".http_response_code()."\n");
+                // NB: can\'t test headers with CLI SAPI
+PHP
+            ,
             [
-                '<?php',
-                'require '.\var_export($file, TRUE).';',
-                'try {',
-                '  ob_start();',
-                '  echo "This should not make it to output";',
-                '  throw new \InvalidArgumentException("Testing handling exceptions");',
-                '} catch (\Exception $e) {',
-                '  '.$class.'::dispatch($e);',
-                '}',
-                '',
-                'fwrite(STDERR, "Code: ".http_response_code()."\n");',
-                '// NB: can\'t test headers with CLI SAPI',
+                'REQUEST_EXCEPTION_DISPATCHER_SOURCE' => $basedir.'/src/ExceptionHandling/RequestExceptionDispatcher.php',
             ]
         );
 
-        $tmpfile = \tempnam(\sys_get_temp_dir(), 'exceptiontest');
-        try {
-            \file_put_contents($tmpfile, $script);
-            $proc = new Process('php '.$tmpfile);
-            $proc->run();
-            $output = $proc->getOutput();
-            $this->assertNotContains('This should not make it to output', $output);
-            $this->assertContains('<div class="error-panel"', $output);
-            $this->assertStringStartsWith('<!DOCTYPE', $output);
-            $this->assertStringEndsWith('</html>', \trim($output));
-            $this->assertSame("Code: 500\n", $proc->getErrorOutput(), 'Should set HTTP status');
-        } finally {
-            \unlink($tmpfile);
-        }
+        $output = $proc->getOutput();
+        $this->assertStringNotContainsString('This should not make it to output', $output);
+        $this->assertStringContainsString('<div class="error-panel"', $output);
+        $this->assertStringStartsWith('<!DOCTYPE', $output);
+        $this->assertStringEndsWith('</html>', \trim($output));
+
+        $error_output       = $proc->getErrorOutput();
+        $error_output_lines = explode("\n", $error_output);
+        $log                = JSON::decodeArray($error_output_lines[0]);
+        $this->assertSame('EMERGENCY', $log['severity']);
+        $this->assertStringContainsString('No container class Dependencies', $log['message']);
+
+        $this->assertSame("Code: 500", $error_output_lines[2], 'Should set HTTP status');
+    }
+
+    public function test_its_static_constructor_creates_instance_that_works_with_default_stackdriver_logger_if_initialised(
+    )
+    {
+        $basedir = \realpath(__DIR__.'/../../../');
+        $proc    = $this->runTemporaryPhpScript(
+            <<<'PHP'
+            use Ingenerator\KohanaExtras\ExceptionHandling\RequestExceptionDispatcher;
+            use Ingenerator\PHPUtils\Logging\StackdriverApplicationLogger;
+            require REQUEST_EXCEPTION_DISPATCHER_SOURCE;
+            require COMPOSER_AUTOLOADER;
+            StackdriverApplicationLogger::initialise(function () {
+                return new StackdriverApplicationLogger(
+                'php://stderr',
+                ['custom_context' => 'from-test']
+                );
+            });
+            RequestExceptionDispatcher::dispatch(new \BadMethodCallException('Anything happened'));            
+PHP
+            ,
+            [
+                'REQUEST_EXCEPTION_DISPATCHER_SOURCE' => $basedir.'/src/ExceptionHandling/RequestExceptionDispatcher.php',
+                'COMPOSER_AUTOLOADER'                 => $basedir.'/vendor/autoload.php',
+            ]
+        );
+
+        $output = $proc->getOutput();
+        $this->assertStringContainsString('<div class="error-panel"', $output);
+        $this->assertStringStartsWith('<!DOCTYPE', $output);
+        $this->assertStringEndsWith('</html>', \trim($output));
+
+        $error_output       = $proc->getErrorOutput();
+        $error_output_lines = explode("\n", $error_output);
+        $log                = JSON::decodeArray($error_output_lines[0]);
+        $this->assertSame('EMERGENCY', $log['severity']);
+        $this->assertStringContainsString('No container class Dependencies', $log['message']);
+        $this->assertSame('from-test', $log['custom_context']);
     }
 
     public function provider_handler_map()
@@ -208,36 +259,66 @@ class RequestExceptionDispatcherTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals($expect, $response->body());
     }
 
-    public function test_it_logs_handler_exceptions_to_syslog_and_returns_generic_error()
+    public function test_it_logs_handler_exceptions_and_returns_generic_error()
     {
-        $logged = $this->withFakeSyslog(
-            function () {
-                $this->default_handler = new ThrowingHandler('The handler broke');
-                $response              = $this->newSubject()->handle(new \InvalidArgumentException('The app broke'));
+        $handler_e             = new \RuntimeException('The handler broke');
+        $this->logger          = new TestLogger;
+        $this->default_handler = new class($handler_e) implements ExceptionHandler {
+            private $e;
+
+            public function __construct($e) { $this->e = $e; }
+
+            public function handle(\Throwable $e): ?\Response
+            {
+                throw $this->e;
+            }
+        };
+
+        $app_e = new \InvalidArgumentException('The app broke');
+
+        $response = $this->newSubject()->handle($app_e);
+        $this->assertGenericErrorResponse($response);
+        $this->assertSame(
+            [
+                [
+                    'level'   => LogLevel::EMERGENCY,
+                    'message' => '[RuntimeException] The handler broke ('.__FILE__.':'.$handler_e->getLine().')',
+                    'context' => ['exception' => $handler_e],
+                ],
+                [
+                    'level'   => LogLevel::EMERGENCY,
+                    'message' => '[InvalidArgumentException] The app broke ('.__FILE__.':'.$app_e->getLine().')',
+                    'context' => ['exception' => $app_e],
+                ],
+            ],
+            $this->logger->records
+        );
+    }
+
+    public function test_it_logs_and_returns_generic_error_with_missing_dependencies()
+    {
+        $exception = new \InvalidArgumentException('I broke');
+        $logged    = $this->withFakeFallbackLogger(
+            function () use ($exception) {
+                $response = RequestExceptionDispatcher::factoryFallback()->handle($exception);
                 $this->assertGenericErrorResponse($response);
             }
         );
 
-        $this->assertContains('The app broke', $logged);
-        $this->assertContains('The handler broke', $logged);
-    }
-
-    public function test_it_logs_to_syslog_and_returns_generic_error_with_missing_dependencies()
-    {
-        $logged = $this->withFakeSyslog(
-            function () {
-                $response = RequestExceptionDispatcher::factoryFallback()->handle(
-                    new \InvalidArgumentException('I broke')
-                );
-                $this->assertGenericErrorResponse($response);
-            }
+        $this->assertSame(
+            [
+                [
+                    'msg' => '[InvalidArgumentException] I broke ('.__FILE__.':'.$exception->getLine().')',
+                    'ctx' => ['exception' => $exception],
+                ],
+            ],
+            $logged
         );
-        $this->assertContains('I broke', $logged);
     }
 
-    public function test_it_logs_to_syslog_and_returns_generic_error_when_forced_to_create_fallback_handler()
+    public function test_it_logs_and_returns_generic_error_when_forced_to_create_fallback_handler()
     {
-        $logged = $this->withFakeSyslog(
+        $logged = $this->withFakeFallbackLogger(
             function () {
                 $response = RequestExceptionDispatcher::factory('No\Such\Class', 'anything', 'anything')
                     ->handle(new \LogicException('No logic!'));
@@ -245,8 +326,9 @@ class RequestExceptionDispatcherTest extends \PHPUnit\Framework\TestCase
             }
         );
 
-        $this->assertContains('No\Such\Class', $logged);
-        $this->assertContains('No logic!', $logged);
+        $this->assertCount(2, $logged, 'Expected two log messages');
+        $this->assertContains('No\Such\Class', $logged[0]['msg']);
+        $this->assertContains('No logic!', $logged[1]['msg']);
     }
 
 
@@ -276,16 +358,47 @@ class RequestExceptionDispatcherTest extends \PHPUnit\Framework\TestCase
         );
     }
 
-    public function setUp()
+    /**
+     * @param string $script
+     * @param array  $vars
+     *
+     * @return Process
+     */
+    protected function runTemporaryPhpScript(string $script, array $vars): Process
+    {
+        $script = \strtr(
+            $script,
+            array_map(function ($var) { return \var_export($var, TRUE); }, $vars)
+        );
+
+        $tmpfile = \tempnam(\sys_get_temp_dir(), 'exceptiontest');
+        try {
+            \file_put_contents($tmpfile, "<?php\n".$script);
+            $proc = new Process('php '.$tmpfile);
+            $proc->run();
+        } finally {
+            \unlink($tmpfile);
+        }
+
+        return $proc;
+    }
+
+    protected function setUp()
     {
         parent::setUp();
         $this->default_handler = new DummyHandler('default');
         $this->dependencies    = NULL;
+        $this->logger          = new NullLogger;
     }
 
     protected function newSubject()
     {
-        return new RequestExceptionDispatcher($this->default_handler, $this->dependencies, $this->handler_map);
+        return new RequestExceptionDispatcher(
+            $this->default_handler,
+            $this->dependencies,
+            $this->logger,
+            $this->handler_map
+        );
     }
 
     /**
@@ -293,20 +406,30 @@ class RequestExceptionDispatcherTest extends \PHPUnit\Framework\TestCase
      *
      * @return array
      */
-    protected function withFakeSyslog($fn)
+    protected function withFakeFallbackLogger($fn)
     {
-        $logged = [];
-        try {
-            RequestExceptionDispatcher::$syslog_func = function ($priority, $message) use (&$logged) {
-                $logged[] = $priority.' - '.$message;
-            };
+        $logger            = new class {
+            public static $messages = [];
 
+            public static function instance() { return new static; }
+
+            public function emergency($message, array $context = [])
+            {
+                static::$messages[] = ['msg' => $message, 'ctx' => $context];
+            }
+        };
+        $logger::$messages = [];
+        $old_fallback_log  = RequestExceptionDispatcher::$fallbackLoggerFactories;
+        try {
+            RequestExceptionDispatcher::$fallbackLoggerFactories = [
+                [\get_class($logger), 'instance'],
+            ];
             $fn();
         } finally {
-            RequestExceptionDispatcher::$syslog_func = 'syslog';
+            RequestExceptionDispatcher::$fallbackLoggerFactories = $old_fallback_log;
         }
 
-        return \implode("\n", $logged);
+        return $logger::$messages;
     }
 
     protected function assertGenericErrorResponse($value)
@@ -330,29 +453,13 @@ class DummyHandler implements ExceptionHandler
         $this->response_prefix = $response_prefix;
     }
 
-    public function handle($e)
+    public function handle(\Throwable $e): ?\Response
     {
         if ($this->response_prefix) {
             return \Response::factory()->body($this->response_prefix.': '.$e->getMessage());
         }
 
         return NULL;
-    }
-}
-
-class ThrowingHandler implements ExceptionHandler
-{
-
-    /**
-     * @var string
-     */
-    protected $msg;
-
-    public function __construct($msg = 'I broke') { $this->msg = $msg; }
-
-    public function handle($e)
-    {
-        throw new \RuntimeException($this->msg, 0, $e);
     }
 }